requirements.txt
discord.py>=2.5.0
flask>=3.1.0
python-dotenv>=1.1.0
requests==2.24.0
Pillow==8.4.0
